Mark integration tests explicitly as such.
