+++ sqlserver/tests/test_connection.py
@@ -381,7 +381,7 @@ def test_connection_failure(aggregator, dd_run_check, instance_docker):
     )
 
 
-@pytest.mark.unit
+@pytest.mark.integration
 @pytest.mark.parametrize(
     "test_case_name,instance_overrides,expected_error_patterns,expected_error",
     [
