+++ yarn/tests/test_integration.py
@@ -1,10 +1,12 @@
 # (C) Datadog, Inc. 2019-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
+import os
 
 import pytest
 
 from datadog_checks.base import AgentCheck
+from datadog_checks.yarn import YarnCheck
 
 from . import common
 
@@ -30,3 +32,25 @@ def assert_check(aggregator):
         aggregator.assert_metric_has_tag(metric, common.LEGACY_CLUSTER_TAG)
 
     aggregator.assert_all_metrics_covered()
+
+
+@pytest.mark.integration
+def test_metadata(aggregator, instance, datadog_agent):
+    check = YarnCheck("yarn", {}, [instance])
+    check.check_id = "test:123"
+
+    check.check(instance)
+
+    raw_version = os.getenv("YARN_VERSION")
+
+    major, minor, patch = raw_version.split(".")
+
+    version_metadata = {
+        "version.scheme": "semver",
+        "version.major": major,
+        "version.minor": minor,
+        "version.patch": patch,
+        "version.raw": raw_version,
+    }
+
+    datadog_agent.assert_metadata("test:123", version_metadata)
